refactor conn pool functions into sculpt_pool.c
